Persist dirty document records
Before this change, an upload failure would be recorded as a retry
job, but the job queue was not intended to provide complete
persistency, and it was possible that the reference to a dirty
document would be lost. Moreover, a retry job was queued only if
an upload failure was reported, which makes the process vulnerable
to unexpected halt of the process.
Now dirty documents will be recorded to a reliable database,
PostgreSQL before they are created, fixing both of those two
problems.