Instantiate tuple objects after they are persisted
A seed data to be persisted may not contain properties such as unique id,
and in the past tuple objects accepted such incomplete state. That required
references to those properties to be guarded with conditionals, where they
are often unnecessary. Those properties were also simply writable because
insertion methods of the repository modified given tuple objects to
complete them, and they were vulnerable to unintended modifications.
That behavior of insertion methods was also inconsistent with one of
query methods.

This fixes the situation by requiring that an object specialized to
represent a new tuple to be provided to insertion methods, and that the
methods instantiate and return common tuple objects.